requirements.txt
pyserial
numpy<1.24